pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
django-debug-toolbar = "^6.0.0"
django-elasticsearch-dsl = "^9.0"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
python-decouple = "^3.8"
dj-database-url = "^3.0.1"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
